Tags can now be sorted by on the News Section and are Add-able on editing a post.

+++ app/Http/Controllers/Backend/CategoriesController.php
@@ -100,6 +100,12 @@ class CategoriesController extends BackendController
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0)
+        {
+            session()->flash('warning', 'Category cannot be deleted as it has posts categorized under it.');
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'The Category was Deleted Successfully!');
